feat: add maegin charts to deal-ticket fee details
