+++ libs/deal-ticket/src/components/deal-ticket/deal-ticket.tsx
@@ -31,7 +31,7 @@ import {
 } from '@vegaprotocol/positions';
 import { toBigNum, removeDecimal } from '@vegaprotocol/utils';
 import { activeOrdersProvider } from '@vegaprotocol/orders';
-import { useEstimateFees } from '../../hooks/use-fee-deal-ticket-details';
+import { useEstimateFees } from '../../hooks/use-estimate-fees';
 import { getDerivedPrice } from '../../utils/get-price';
 import type { OrderInfo } from '@vegaprotocol/types';
 
@@ -54,8 +54,6 @@ import {
 import { OrderTimeInForce, OrderType } from '@vegaprotocol/types';
 import { useOrderForm } from '../../hooks/use-order-form';
 import { useDataProvider } from '@vegaprotocol/data-provider';
-
-import { marketMarginDataProvider } from '@vegaprotocol/positions';
 
 export interface DealTicketProps {
   market: Market;
@@ -174,12 +172,6 @@ export const DealTicket = ({
 
   const assetSymbol =
     market.tradableInstrument.instrument.product.settlementAsset.symbol;
-
-  const { data: currentMargins } = useDataProvider({
-    dataProvider: marketMarginDataProvider,
-    variables: { marketId: market.id, partyId: pubKey || '' },
-    skip: !pubKey,
-  });
 
   useEffect(() => {
     if (!pubKey) {
@@ -492,8 +484,6 @@ export const DealTicket = ({
           generalAccountBalance={generalAccountBalance}
           positionEstimate={positionEstimate?.estimatePosition}
           market={market}
-          currentInitialMargin={currentMargins?.initialLevel}
-          currentMaintenanceMargin={currentMargins?.maintenanceLevel}
         />
       </form>
     </TinyScroll>
