+++ libs/deal-ticket/src/hooks/index.ts
@@ -1,2 +1,2 @@
 export * from './__generated__/EstimateOrder';
-export * from './use-fee-deal-ticket-details';
+export * from './use-estimate-fees';
